apps/OOBE/requirements.txt
Pillow==8.1.2
svgwrite==1.4.1
cairosvg==2.5.2
